chore: watcher cyclical dependencies
