+++ packages/plugin-core/src/_extension.ts
@@ -95,7 +95,7 @@ import {
 } from "./workspace";
 import { DendronCodeWorkspace } from "./workspace/codeWorkspace";
 import { DendronNativeWorkspace } from "./workspace/nativeWorkspace";
-import { WorkspaceInitFactory } from "./workspace/workspaceInitializer";
+import { WorkspaceInitFactory } from "./workspace/WorkspaceInitFactory";
 import { WSUtils } from "./WSUtils";
 
 const MARKDOWN_WORD_PATTERN = new RegExp("([\\w\\.\\#]+)");
