+++ packages/plugin-core/src/commands/CreateDailyJournal.ts
@@ -1,26 +1,24 @@
 import { ConfigUtils, VaultUtils } from "@dendronhq/common-all";
 import _ from "lodash";
 import { DENDRON_COMMANDS } from "../constants";
+import { IDendronExtension } from "../dendronExtensionInterface";
 import { JournalNote } from "../traits/journal";
-import { getDWorkspace } from "../workspace";
 import {
   CommandOpts,
   CreateNoteWithTraitCommand,
 } from "./CreateNoteWithTraitCommand";
 
 export class CreateDailyJournalCommand extends CreateNoteWithTraitCommand {
-  constructor() {
-    super("dendron.journal", new JournalNote());
+  constructor(ext: IDendronExtension) {
+    super(ext, "dendron.journal", new JournalNote());
 
     // override the key to maintain compatibility
     this.key = DENDRON_COMMANDS.CREATE_DAILY_JOURNAL_NOTE.key;
   }
 
   override execute(opts: CommandOpts): Promise<void> {
-    const config = getDWorkspace().config;
+    const config = this._extension.getDWorkspace().config;
     const journalConfig = ConfigUtils.getJournal(config);
-
-    const { engine } = getDWorkspace();
 
     if (_.isUndefined(journalConfig.dailyVault)) {
       return super.execute({ ...opts });
@@ -28,7 +26,7 @@ export class CreateDailyJournalCommand extends CreateNoteWithTraitCommand {
       const dailyVault = journalConfig.dailyVault;
       const vault = dailyVault
         ? VaultUtils.getVaultByName({
-            vaults: engine.vaults,
+            vaults: this._extension.getEngine().vaults,
             vname: dailyVault,
           })
         : undefined;
