+++ packages/plugin-core/src/commands/MoveHeader.ts
@@ -29,10 +29,8 @@ import _ from "lodash";
 import path from "path";
 import { Location } from "vscode";
 import { DendronQuickPickerV2 } from "../components/lookup/types";
-import {
-  NoteLookupProviderUtils,
-  PickerUtilsV2,
-} from "../components/lookup/utils";
+import { PickerUtilsV2 } from "../components/lookup/utils";
+import { NoteLookupProviderUtils } from "../components/lookup/NotePickerUtils";
 import { DENDRON_COMMANDS } from "../constants";
 import { delayedUpdateDecorations } from "../features/windowDecorations";
 import { VSCodeUtils } from "../vsCodeUtils";
