+++ packages/plugin-core/src/workspace/workspaceInitializer.ts
@@ -1,10 +1,5 @@
 import { DVault, DWorkspaceV2 } from "@dendronhq/common-all";
 import { WorkspaceService } from "@dendronhq/engine-server";
-import { WORKSPACE_ACTIVATION_CONTEXT } from "../constants";
-import { StateService } from "../services/stateService";
-import { BlankInitializer } from "./blankInitializer";
-import { SeedBrowserInitializer } from "./seedBrowserInitializer";
-import { TutorialInitializer } from "./tutorialInitializer";
 
 /**
  * Type that can execute custom code as part of workspace creation and opening of a workspace.
@@ -29,21 +24,3 @@ export type WorkspaceInitializer = {
    */
   onWorkspaceOpen?(opts: { ws: DWorkspaceV2 }): Promise<void>;
 };
-
-/**
- * Factory class for creating WorkspaceInitializer types
- */
-export class WorkspaceInitFactory {
-  static create(): WorkspaceInitializer | undefined {
-    switch (StateService.instance().getActivationContext()) {
-      case WORKSPACE_ACTIVATION_CONTEXT.TUTORIAL:
-        return new TutorialInitializer();
-
-      case WORKSPACE_ACTIVATION_CONTEXT.SEED_BROWSER:
-        return new SeedBrowserInitializer();
-
-      default:
-        return new BlankInitializer();
-    }
-  }
-}
